Delat upp Customer helt
