+++ CustomerUserList.cs
@@ -1,0 +1,21 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace DotNetDynamos
+{
+    internal partial class Customer : AllUsers
+    {
+        public override void UserList()
+
+        {
+            Console.WriteLine("Customer Users:");
+            foreach (KeyValuePair<int, string> customerUser in CustomerUsers)
+            {
+                Console.WriteLine($"ID: {customerUser.Key}, Username: {customerUser.Value}");
+            }
+        }
+    }
+}
